==> 02_generate_case_study_files.ipynb <==
# Generate simulated data to link

In this case study, we imagine running a person linkage on a simulated 2030 Census Unedited File (CUF) --
see the main notebook for more details, including references used throughout this notebook.
This notebook creates simulated input (CUF) and reference files approximating what would be used in such a process,
emulating the methods found in publicly available descriptions of the Census Bureau's primary person linkage system, PVS.

# Query planning is now on by default, but it has some rough edges.
# See https://github.com/dask/dask/issues/10995 for general discussion
# and https://github.com/dask/dask-expr/issues/1060 for the particular
# issue I ran into.
try:
    import dask

    dask.config.set({"dataframe.query-planning": False})
except ImportError:
    pass

import os, time
import warnings
import datetime

# Importing pandas for access, regardless of whether we are using it as the compute engine
import pandas
import numpy
from IPython.display import display

%load_ext autoreload
%autoreload 1

from person_linkage_case_study_utils import distributed_compute, utils

warnings.simplefilter(action="ignore", category=FutureWarning)

# DO NOT EDIT if this notebook is not called 02_generate_case_study_files.ipynb!
# This notebook is designed to be run with papermill; this cell is tagged 'parameters'
data_to_use = "small_sample"
input_dir = "output/01_generate_pseudopeople_simulated_datasets"
output_dir = "output/02_generate_case_study_files"
compute_engine = "pandas"
compute_engine_num_workers = 10
compute_engine_cpus_per_worker = 2
compute_engine_threads_per_worker = 1
compute_engine_memory_per_worker = "10GB"
queue = None
account = None
# NOTE: This is, as Dask requests, a directory local to the compute node.
# But IHME's cluster doesn't support this very well -- it can be small-ish,
# full of stuff from other users, etc.
compute_engine_local_directory = (
    f"/tmp/{os.environ['USER']}_{int(time.time())}_person_linkage_case_study"
)
compute_engine_log_directory = f"{output_dir}/{data_to_use}/logs"
walltime = None
compute_engine_memory_constrained = True
scheduler = "slurm"

if compute_engine != "pandas":
    utils.ensure_empty(compute_engine_local_directory)

print(datetime.datetime.now())

df_ops, pd = distributed_compute.start_compute_engine(
    compute_engine,
    num_workers=compute_engine_num_workers,
    cpus_per_worker=compute_engine_cpus_per_worker,
    threads_per_worker=compute_engine_threads_per_worker,
    memory_per_worker=compute_engine_memory_per_worker,
    worker_walltime=walltime,
    local_directory=compute_engine_local_directory,
    log_directory=compute_engine_log_directory,
    memory_constrained=compute_engine_memory_constrained,
    scheduler=scheduler,
    queue=queue,
    account=account,
)

def save_file_with_ground_truth(file_name, file, ground_truth):
    # Check that file and ground truth have the same records
    assert (
        # record_id is unique
        len(file)
        == len(df_ops.drop_duplicates(file[["record_id"]]))
    )
    assert len(ground_truth) == len(
        file[["record_id"]].merge(
            ground_truth[["record_id"]], on="record_id", how="inner"
        )
    )

    file_path = f"{output_dir}/{data_to_use}/{file_name}.parquet"
    utils.remove_path(file_path)
    file.to_parquet(file_path)

    ground_truth_path = f"{output_dir}/{data_to_use}/{file_name}_ground_truth.parquet"
    utils.remove_path(ground_truth_path)
    ground_truth.to_parquet(ground_truth_path)

## Load pseudopeople simulated datasets

### Record ID tracking (data lineage)

We do a little bit of work here to enable tracking the "ground truth" (the simulant IDs from
pseudopeople).
We give each pseudopeople record/row a unique identifier for tracking, and then we immediately
separate the ground truth information (which we would not have if we were using real data)
from the rest of the columns (which we would have).
The ground truth is only used in the specific "ground truth" section of this notebook,
to help avoid accidentally leaking information into the case study.

Since we also combine/aggregate pseudopeople records as part of the process of generating the
simulated reference files, ground truth is a bit more complicated than you might imagine.
For example, the ground truth may tell us that a single row in a reference file is actually
a composite of several individuals, because even the deterministic linkage (by SSN) we use
here is not without error.

We handle this by tracking *all* source records used in the construction of each record in
our reference files.
This is achieved by having a table mapping composite record IDs to the "source record IDs"
(IDs of records that were directly generated by pseudopeople).
When we aggregate records, this is combined accordingly.

def add_unique_id_col(df, col_name="unique_id", value_prefix=""):
    if compute_engine == "pandas" or compute_engine.startswith("modin"):
        return (
            df.reset_index()
            .rename(columns={"index": col_name})
            .assign(**{col_name: lambda df: value_prefix + df[col_name].astype(str)})
        )
    elif compute_engine.startswith("dask"):
        # Can use cumsum as in https://stackoverflow.com/a/60852409/ if it needs
        # to be incrementing, but we just need uniqueness

        def add_id_to_partition(df_part, partition_info=None):
            return df_part.assign(**{col_name: range(len(df_part))}).assign(
                **{
                    col_name: lambda x: (
                        value_prefix
                        + str(
                            partition_info["number"]
                            if partition_info is not None
                            else 0
                        )
                        + "_"
                        + x[col_name].astype(str)
                    ).astype("large_string[pyarrow]")
                }
            )

        df = df.map_partitions(add_id_to_partition)

        return df_ops.ensure_large_string_capacity(df)
    else:
        raise ValueError()


def add_unique_record_id(df, dataset_name):
    return add_unique_id_col(df, col_name="record_id", value_prefix=f"{dataset_name}_")


# Initializes a table listing the pairs between record_ids and source record_ids.
# Should only be called on "source records"; that is, records that
# come directly out of pseudopeople.
def record_id_to_single_source_record_pairs(df, source_col="record_id"):
    if source_col == "record_id":
        # We can't have duplicate column names, so we make a new column
        # literally called 'source_col'
        df = df.assign(source_col=lambda df: df[source_col])
        source_col = "source_col"

    return df[["record_id", source_col]].rename(
        columns={source_col: "source_record_id"}
    )

# Operations that aggregate records, combining the source record pairs
# between all records that are aggregated


def merge_preserving_source_records(
    dfs, source_record_pairings, new_record_id_prefix, *args, **kwargs
):
    assert len(dfs) == len(source_record_pairings)
    for df in dfs:
        assert "record_id" in df.columns

    on = kwargs.get("on", None)
    if on is not None:
        # If there are nulls in any of the merge columns, they can't match to anything
        dfs = [df.dropna(subset=on, how="any") for df in dfs]

    result = dfs[0]
    source_record_pairs = source_record_pairings[0]
    dfs_and_source_record_pairs_to_combine = list(
        zip(dfs[1:], source_record_pairings[1:])
    )
    for index, (df_to_merge, source_record_pairs_to_merge) in enumerate(
        dfs_and_source_record_pairs_to_combine
    ):
        result = result.merge(df_to_merge, *args, **kwargs)
        if index == len(dfs_and_source_record_pairs_to_combine) - 1:
            # Since this is the last step, these are the record_ids that will actually be returned
            accumulate_step_record_id_prefix = new_record_id_prefix
        else:
            # A dummy intermediate -- this shouldn't be exposed to the user
            accumulate_step_record_id_prefix = f"merge_iter_{index}"

        result = df_ops.rebalance(
            add_unique_record_id(result, accumulate_step_record_id_prefix)
        )
        source_record_pairs = df_ops.concat(
            [
                # The pairs that were already in result
                source_record_pairs.rename(columns={"record_id": "record_id_x"})
                .merge(
                    result[["record_id", "record_id_x"]].dropna(how="any"),
                    on="record_id_x",
                )
                .drop(columns=["record_id_x"]),
                # The new ones
                source_record_pairs_to_merge.rename(
                    columns={"record_id": "record_id_y"}
                )
                .merge(
                    result[["record_id", "record_id_y"]].dropna(how="any"),
                    on="record_id_y",
                )
                .drop(columns=["record_id_y"]),
            ]
        )
        result = result.drop(columns=["record_id_x", "record_id_y"])

    return result, source_record_pairs


def dedupe_preserving_source_records(
    df, source_record_pairs, columns_to_dedupe, new_record_id_prefix
):
    result = df_ops.drop_duplicates(df[columns_to_dedupe])
    result = add_unique_record_id(result, new_record_id_prefix)
    df_to_result_mapping = (
        df[["record_id"] + columns_to_dedupe]
        .rename(columns={"record_id": "record_id_pre_dedupe"})
        .merge(result, on=columns_to_dedupe)[["record_id", "record_id_pre_dedupe"]]
    )
    result_source_record_pairs = (
        source_record_pairs.rename(columns={"record_id": "record_id_pre_dedupe"})
        .merge(df_to_result_mapping, on="record_id_pre_dedupe")
        .drop(columns=["record_id_pre_dedupe"])
    )
    return result, result_source_record_pairs


def concat_preserving_source_records(dfs, source_record_pairings, new_record_id_prefix):
    dfs = [df.rename(columns={"record_id": "record_id_pre_concat"}) for df in dfs]
    result = df_ops.concat(dfs, ignore_index=True)
    result = add_unique_record_id(result, new_record_id_prefix)

    record_id_mapping = result[["record_id", "record_id_pre_concat"]]
    validate_kwarg = (
        {
            "validate": "m:1",
        }
        if not compute_engine.startswith("dask")
        else {}
    )  # Dask doesn't support validate

    all_source_record_pairings = df_ops.persist(
        df_ops.concat(source_record_pairings, ignore_index=False)
    )
    result_source_record_pairs = df_ops.persist(
        all_source_record_pairings.rename(columns={"record_id": "record_id_pre_concat"})
        .merge(record_id_mapping, on="record_id_pre_concat", **validate_kwarg)
        .drop(columns=["record_id_pre_concat"])
    )

    if compute_engine.startswith("dask"):
        # Manual alternative to the 'validate' kwarg
        assert len(result_source_record_pairs) == len(all_source_record_pairings)

    return result.drop(columns=["record_id_pre_concat"]), result_source_record_pairs

### Custom noise types

A few noise types are really important here, but not implemented in pseudopeople, so we implement them as "add-ons."

def split_columns_incorrectly(
    df: pandas.DataFrame, columns: list[str], row_probability: float, seed: int = 1234
):
    rng = numpy.random.default_rng(seed)

    to_split_incorrectly = pandas.Series(
        rng.random(len(df)) < row_probability, index=df.index
    )
    full_name = (
        df.loc[to_split_incorrectly, columns]
        .fillna("")
        .agg(" ".join, axis=1)
        .str.strip()
    )

    num_spaces_or_hyphens = full_name.str.count("[ -]")
    split_name = full_name.str.split("[ -]", regex=True, expand=True).fillna("").values

    split_points = numpy.ceil(
        numpy.sort(rng.random((len(full_name), len(columns) - 1)), axis=1)
        * num_spaces_or_hyphens.values.reshape((len(full_name), 1))
    ).astype(int)
    string_start_points_1d = (
        numpy.arange(split_name.shape[0], dtype=int) * split_name.shape[1]
    )
    string_end_points_1d = numpy.append(
        string_start_points_1d[1:], [split_name.shape[0] * split_name.shape[1]]
    )
    last_split_points_1d = None

    for split_index, column_name in enumerate(columns):
        # Map into 1D space
        new_column = split_name.copy().reshape(-1)

        if last_split_points_1d is not None:
            erase_before_slices = tuple(
                slice(string_start_point, last_split_point)
                for string_start_point, last_split_point in zip(
                    string_start_points_1d, last_split_points_1d
                )
            )
        else:
            erase_before_slices = tuple()

        if split_index < split_points.shape[1]:
            split_points_1d = (
                numpy.arange(split_name.shape[0], dtype=int) * split_name.shape[1]
                + split_points[:, split_index]
            )
            erase_after_slices = tuple(
                slice(split_point, string_end_point)
                for split_point, string_end_point in zip(
                    split_points_1d, string_end_points_1d
                )
            )
        else:
            erase_after_slices = tuple()

        new_column[numpy.r_[erase_before_slices + erase_after_slices]] = ""
        new_column[new_column != ""] += " "
        df.loc[to_split_incorrectly, column_name] = (
            new_column.reshape(split_name.shape)
        ).sum(axis=1)
        df.loc[to_split_incorrectly, column_name] = df.loc[
            to_split_incorrectly, column_name
        ].str.strip()

        last_split_points_1d = split_points_1d

    return df

def shuffle_columns(
    df: pandas.DataFrame, columns: list[str], row_probability: float, seed: int = 1234
):
    rng = numpy.random.default_rng(seed)

    to_shuffle = pandas.Series(rng.random(len(df)) < row_probability, index=df.index)
    original_values = df.loc[to_shuffle, columns].values
    random_ordering = numpy.argsort(rng.random(original_values.shape), axis=1)
    df.loc[to_shuffle, columns] = numpy.take_along_axis(
        original_values, random_ordering, axis=1
    )

    return df

def truncate_column(
    df: pandas.DataFrame,
    column: str,
    cell_probability: float,
    truncation_length: int,
    seed: int = 1234,
):
    rng = numpy.random.default_rng(seed)
    to_truncate = pandas.Series(rng.random(len(df)) < cell_probability, index=df.index)
    df.loc[to_truncate, "last_name"] = df.loc[to_truncate, "last_name"].str[
        :truncation_length
    ]
    return df

def apply_custom_noise_type(df, noise_type, *args, seed: int = 1234, **kwargs):
    if compute_engine.startswith("dask"):
        return df.map_partitions(
            lambda partition, partition_info=None: noise_type(
                partition,
                *args,
                seed=(
                    seed
                    + (partition_info["number"] if partition_info is not None else 1)
                    * 10_000
                ),
                **kwargs,
            ),
            meta=df,
        )
    else:
        return noise_type(
            df,
            *args,
            seed=seed,
            **kwargs,
        )

### 1040 tax filings

We assume that the last 5 years of taxes would be available and used in the construction of the reference files -- see section about reference files below.

Note that these are retrieved by *tax* year, so the 2029 taxes would be available in early 2030
(around when our hypothetical case study is taking place).

tax_years = list(range(2025, 2030))
tax_years

%%time

# Combine 1040 for all years.
simulated_taxes_1040 = df_ops.concat(
    [
        df_ops.read_parquet(
            f"{input_dir}/{data_to_use}/simulated_taxes_1040_{year}.parquet",
            columns=[
                "simulant_id",
                "ssn",
                "first_name",
                "middle_initial",
                "last_name",
                "mailing_address_street_number",
                "mailing_address_street_name",
                "mailing_address_unit_number",
                "mailing_address_po_box",
                "mailing_address_city",
                "mailing_address_state",
                "mailing_address_zipcode",
            ],
        )
        for year in tax_years
    ],
    ignore_index=True,
)

simulated_taxes_1040 = apply_custom_noise_type(
    simulated_taxes_1040,
    split_columns_incorrectly,
    ["first_name", "middle_initial", "last_name"],
    row_probability=0.01,
    seed=1,
)

# "... many of the [IRS] records contain only the first four letters of the last name."
# (Brown et al. 2023, p.30, footnote 19)
# Note that this truncation only matters for ITIN PIKing since for SSNs that are present in SSA we use name from SSA.
PROPORTION_OF_IRS_RECORDS_WITH_TRUNCATION = (
    0.4  # is this a good guess at "many" in the quote above?
)
simulated_taxes_1040 = apply_custom_noise_type(
    simulated_taxes_1040,
    truncate_column,
    "last_name",
    cell_probability=PROPORTION_OF_IRS_RECORDS_WITH_TRUNCATION,
    truncation_length=4,
    seed=2,
)

simulated_taxes_1040 = apply_custom_noise_type(
    simulated_taxes_1040,
    shuffle_columns,
    ["first_name", "middle_initial", "last_name"],
    row_probability=0.01,
    seed=3,
)

simulated_taxes_1040 = df_ops.persist(
    add_unique_record_id(simulated_taxes_1040, "simulated_1040")
)
simulated_taxes_1040_source_record_pairs = df_ops.persist(
    record_id_to_single_source_record_pairs(simulated_taxes_1040)
)

simulated_taxes_1040_ground_truth = simulated_taxes_1040[["record_id", "simulant_id"]]
simulated_taxes_1040 = simulated_taxes_1040.drop(columns=["simulant_id"])
simulated_taxes_1040, simulated_taxes_1040_ground_truth = df_ops.persist(
    simulated_taxes_1040, simulated_taxes_1040_ground_truth
)
simulated_taxes_1040

### W2/1099 tax filings

We assume that the last 5 years of taxes would be available and used in the construction of the reference files --
see section about reference files below.

Note that these are retrieved by *tax* year, so the 2029 taxes would be available in early 2030
(around when our hypothetical case study is taking place).

%%time

# Combine W2/1099 for all years.
simulated_w2_1099 = df_ops.concat(
    [
        df_ops.read_parquet(
            f"{input_dir}/{data_to_use}/simulated_taxes_w2_and_1099_{year}.parquet",
            columns=[
                "simulant_id",
                "ssn",
                "first_name",
                "middle_initial",
                "last_name",
                "mailing_address_street_number",
                "mailing_address_street_name",
                "mailing_address_unit_number",
                "mailing_address_po_box",
                "mailing_address_city",
                "mailing_address_state",
                "mailing_address_zipcode",
            ],
        )
        for year in tax_years
    ],
    ignore_index=True,
)

simulated_w2_1099 = apply_custom_noise_type(
    simulated_w2_1099,
    split_columns_incorrectly,
    ["first_name", "middle_initial", "last_name"],
    row_probability=0.01,
    seed=4,
)

simulated_w2_1099 = apply_custom_noise_type(
    simulated_w2_1099,
    truncate_column,
    "last_name",
    cell_probability=PROPORTION_OF_IRS_RECORDS_WITH_TRUNCATION,
    truncation_length=4,
    seed=5,
)

simulated_w2_1099 = apply_custom_noise_type(
    simulated_w2_1099,
    shuffle_columns,
    ["first_name", "middle_initial", "last_name"],
    row_probability=0.01,
    seed=6,
)

simulated_w2_1099 = df_ops.persist(
    add_unique_record_id(simulated_w2_1099, "simulated_w2_1099")
)
simulated_w2_1099_source_record_pairs = df_ops.persist(
    record_id_to_single_source_record_pairs(simulated_w2_1099)
)

simulated_w2_1099_ground_truth = simulated_w2_1099[["record_id", "simulant_id"]]
simulated_w2_1099 = simulated_w2_1099.drop(columns=["simulant_id"])
simulated_w2_1099, simulated_w2_1099_ground_truth = df_ops.persist(
    simulated_w2_1099, simulated_w2_1099_ground_truth
)
simulated_w2_1099

simulated_taxes, simulated_taxes_source_record_pairs = concat_preserving_source_records(
    [simulated_taxes_1040, simulated_w2_1099],
    [simulated_taxes_1040_source_record_pairs, simulated_w2_1099_source_record_pairs],
    new_record_id_prefix="simulated_taxes",
)

simulated_taxes_1040_with_itins = simulated_taxes_1040[
    simulated_taxes_1040.ssn.notnull() & simulated_taxes_1040.ssn.str.startswith("9")
]
simulated_taxes_1040_with_itins_source_record_pairs = (
    simulated_taxes_1040_source_record_pairs.merge(
        simulated_taxes_1040_with_itins[["record_id"]], on="record_id", how="inner"
    )
)
simulated_taxes_1040_with_itins, simulated_taxes_1040_with_itins_source_record_pairs = (
    df_ops.persist(
        simulated_taxes_1040_with_itins,
        simulated_taxes_1040_with_itins_source_record_pairs,
    )
)

%xdel simulated_taxes_1040
%xdel simulated_taxes_1040_source_record_pairs

%xdel simulated_w2_1099
%xdel simulated_w2_1099_source_record_pairs

simulated_taxes, simulated_taxes_source_record_pairs = df_ops.persist(
    simulated_taxes, simulated_taxes_source_record_pairs
)

### 2030 Census Unedited File (CUF)

For now, we gloss over the data schema for addresses.
We don't know how addresses would be formatted in the CUF (and it's hard to guess, because
address is not part of the Census form), but it likely would have some of these fields
(street number, street name, etc) combined.

While PVS input files do not in general have names split into first, middle, and last,
I am guessing the CUF **would** have first name, middle initial, last name (which is how pseudopeople
generates it), because that [matches the Census questionnaire](https://www2.census.gov/programs-surveys/decennial/2020/technical-documentation/questionnaires-and-instructions/questionnaires/2020-informational-questionnaire-english_DI-Q1.pdf).

%%time

simulated_census_2030 = df_ops.read_parquet(
    f"{input_dir}/{data_to_use}/simulated_census_2030.parquet"
)

simulated_census_2030 = apply_custom_noise_type(
    simulated_census_2030,
    split_columns_incorrectly,
    ["first_name", "middle_initial", "last_name"],
    row_probability=0.01,
    seed=7,
)

simulated_census_2030 = apply_custom_noise_type(
    simulated_census_2030,
    shuffle_columns,
    ["first_name", "middle_initial", "last_name"],
    row_probability=0.01,
    seed=8,
)

simulated_census_2030 = df_ops.persist(
    add_unique_record_id(simulated_census_2030, "simulated_census_2030")
)

simulated_census_2030_ground_truth = simulated_census_2030[["record_id", "simulant_id"]]
simulated_census_2030 = simulated_census_2030.drop(columns=["simulant_id"])
simulated_census_2030, simulated_census_2030_ground_truth = df_ops.persist(
    simulated_census_2030, simulated_census_2030_ground_truth
)
simulated_census_2030

save_file_with_ground_truth(
    "simulated_census_2030", simulated_census_2030, simulated_census_2030_ground_truth
)

%xdel simulated_census_2030

### Simulated SSA Numident

%%time

simulated_ssa_numident = df_ops.read_parquet(
    f"{input_dir}/{data_to_use}/simulated_ssa_numident.parquet"
)

simulated_ssa_numident = apply_custom_noise_type(
    simulated_ssa_numident,
    split_columns_incorrectly,
    ["first_name", "middle_name", "last_name"],
    row_probability=0.01,
    seed=9,
)

simulated_ssa_numident = apply_custom_noise_type(
    simulated_ssa_numident,
    shuffle_columns,
    ["first_name", "middle_name", "last_name"],
    row_probability=0.01,
    seed=10,
)

if data_to_use == "ri":
    # Remove simulants who were never observed in Rhode Island -- this is unrealistic to real life!
    # We only do this to create a medium-sized set of reference files for computational testing.
    simulated_ssa_numident = (
        simulated_ssa_numident.merge(
            df_ops.drop_duplicates(
                simulated_taxes_1040_ground_truth[["simulant_id"]]
            ).assign(observed_in_taxes_1040=1),
            on="simulant_id",
            how="left",
        )
        .merge(
            df_ops.drop_duplicates(
                simulated_w2_1099_ground_truth[["simulant_id"]]
            ).assign(observed_in_taxes_w2_1099=1),
            on="simulant_id",
            how="left",
        )
        .merge(
            df_ops.drop_duplicates(
                simulated_census_2030_ground_truth[["simulant_id"]]
            ).assign(observed_in_census_2030=1),
            on="simulant_id",
            how="left",
        )
        .assign(
            observed_ever=lambda df: (df.observed_in_taxes_1040 == 1)
            | (df.observed_in_taxes_w2_1099 == 1)
            | (df.observed_in_census_2030 == 1)
        )
        .pipe(lambda df: df[df.observed_ever])
        .drop(
            columns=[
                "observed_in_taxes_1040",
                "observed_in_taxes_w2_1099",
                "observed_in_census_2030",
                "observed_ever",
            ]
        )
    )

%xdel simulated_census_2030_ground_truth

simulated_ssa_numident = df_ops.persist(
    add_unique_record_id(simulated_ssa_numident, "simulated_ssa_numident")
)
simulated_ssa_numident_source_record_pairs = df_ops.persist(
    record_id_to_single_source_record_pairs(simulated_ssa_numident)
)

simulated_ssa_numident_ground_truth = simulated_ssa_numident[
    ["record_id", "simulant_id"]
]
simulated_ssa_numident = simulated_ssa_numident.drop(columns=["simulant_id"])
simulated_ssa_numident, simulated_ssa_numident_ground_truth = df_ops.persist(
    simulated_ssa_numident, simulated_ssa_numident_ground_truth
)
simulated_ssa_numident

source_record_ground_truth = df_ops.persist(
    df_ops.concat(
        [
            simulated_ssa_numident_ground_truth,
            simulated_taxes_1040_ground_truth,
            simulated_w2_1099_ground_truth,
        ]
    ).rename(columns={"record_id": "source_record_id"})
)
source_record_ground_truth

%xdel simulated_taxes_1040_ground_truth
%xdel simulated_w2_1099_ground_truth

## Create reference files

> The Census Numident – all Social Security Administration (SSA) Numident SSN records are
  edited (collapsed) to produce a Census Numident file that contains “one best-data record” for
  each SSN. All variants of name information for each SSN are retained in the Alternate Name
  Numident file, while all variants of date of birth data are retained in the Alternate DOB
  Numident. The SSN-PIK crosswalk file is used to attach a corresponding unique PIK value for
  each SSN value in the Census Numident file.

### Census Numident

Luque and Wagner, p. 4:
  
> The SSA Numident file contains all transactions ever recorded against any single SSN - with each entry
representing an addition or change (such as name changes) to the SSN record. This file is edited to
create the **Census Numident**, which contains one record for each SSN. Each SSN record in the Census
Numident contains name, DOB, sex, race, place of birth, parents’ name, citizenship status and date of death information.

and in footnote 5:

> Name edits, DOB reconciliation, and race identifiers are some of the edits conducted to produce this Numident
file. **The resulting Numident file contains the most recent name and DOB data.**

We are missing quite a few columns, since they are missing in pseudopeople's SSA Numident: race, place of birth, parents' name,
citizenship status.
However, I'm pretty sure none of these are used in linking.

def fill_dates(df, fill_with):
    return (
        # Replace invalid dates with nans
        pd.to_datetime(df.event_date, format="%Y%m%d", errors="coerce").fillna(
            pandas.to_datetime("2100-01-01" if fill_with == "latest" else "1900-01-01")
        )
    )


def best_data_from_columns(df, columns, best_is_latest=True):
    # We don't want to throw out events with a missing/invalid date, so we'll fill them with the value *least* likely to be chosen
    # (earlier than all values if taking the latest, later than all values if taking the earliest).
    fill_with = "earliest" if best_is_latest else "latest"

    result = (
        df
        # Without mutating the existing date column, get one that is actually
        # a date type and can be used for sorting.
        # Note: we actually convert this to an integer for sorting purposes, because Modin was having trouble
        # sorting by it as an actual datetime
        .assign(
            event_date_for_sort=lambda df: fill_dates(df, fill_with=fill_with).astype(
                numpy.int64
            )
            // 10**9
        )
        .dropna(subset=columns, how="all")
        .pipe(
            lambda df: df_ops.drop_duplicates(
                df,
                subset="ssn",
                sort_col="event_date_for_sort",
                keep=("last" if best_is_latest else "first"),
            )
        )[["record_id", "ssn"] + columns]
    )

    return result, record_id_to_single_source_record_pairs(result)


best_name, best_name_source_record_pairs = best_data_from_columns(
    simulated_ssa_numident,
    columns=["first_name", "middle_name", "last_name"],
)

best_date_of_birth, best_date_of_birth_source_record_pairs = best_data_from_columns(
    simulated_ssa_numident,
    columns=["date_of_birth"],
)

best_date_of_death, best_date_of_death_source_record_pairs = best_data_from_columns(
    simulated_ssa_numident[simulated_ssa_numident.event_type == "death"],
    columns=["event_date"],
)
best_date_of_death = best_date_of_death.rename(columns={"event_date": "date_of_death"})

simulated_census_numident, simulated_census_numident_source_record_pairs = (
    df_ops.persist(
        merge_preserving_source_records(
            [best_name, best_date_of_birth, best_date_of_death],
            [
                best_name_source_record_pairs,
                best_date_of_birth_source_record_pairs,
                best_date_of_death_source_record_pairs,
            ],
            new_record_id_prefix="simulated_census_numident",
            on=["ssn"],
            how="left",
        )
    )
)
simulated_census_numident

### Alternate Name Numident

Wagner and Layne, p. 9:

>  All variants of name information for each SSN are retained in the Alternate Name
Numident file...

(
    simulated_alternate_name_numident,
    simulated_alternate_name_numident_source_record_pairs,
) = df_ops.persist(
    dedupe_preserving_source_records(
        simulated_ssa_numident,
        simulated_ssa_numident_source_record_pairs,
        columns_to_dedupe=["ssn", "first_name", "middle_name", "last_name"],
        new_record_id_prefix="simulated_alternate_name_numident",
    )
)
simulated_alternate_name_numident

df_ops.compute(
    df_ops.groupby_agg_small_groups(
        simulated_alternate_name_numident, by="ssn", agg_func=lambda x: x.size()
    ).describe()
)

# Show alternate names for the same simulated SSN
simulated_alternate_name_numident.merge(
    df_ops.groupby_agg_small_groups(
        simulated_alternate_name_numident[["ssn"]],
        by="ssn",
        agg_func=lambda x: x.size(),
    )
    .pipe(lambda x: x[x > 1])
    .reset_index()[["ssn"]],
    on="ssn",
    how="inner",
).sort_values("ssn").tail(100)

### Alternate DOB Numident

Wagner and Layne, p. 9:

> ... while all variants of date of birth data are retained in the Alternate DOB
Numident.

(
    simulated_alternate_dob_numident,
    simulated_alternate_dob_numident_source_record_pairs,
) = df_ops.persist(
    dedupe_preserving_source_records(
        simulated_ssa_numident,
        simulated_ssa_numident_source_record_pairs,
        columns_to_dedupe=["ssn", "date_of_birth"],
        new_record_id_prefix="simulated_alternate_dob_numident",
    )
)
simulated_alternate_dob_numident

df_ops.compute(
    df_ops.groupby_agg_small_groups(
        simulated_alternate_dob_numident, by="ssn", agg_func=lambda x: x.size()
    ).describe()
)

simulated_alternate_dob_numident.merge(
    df_ops.groupby_agg_small_groups(
        simulated_alternate_dob_numident[["ssn"]], by="ssn", agg_func=lambda x: x.size()
    )
    .pipe(lambda x: x[x > 1])
    .reset_index()[["ssn"]],
    on="ssn",
    how="inner",
).sort_values("ssn").tail(100)

### Name/DOB Reference File

Wagner and Layne, p. 9:

> The Name and DOB Reference files are reformatted versions of the Census Numident
and includes **all possible combinations of alternate names and dates of birth, as well as
ITIN data**. All of the reference files contain SSN/ITIN and the corresponding PIK. When
an input record is linked to a reference file, the corresponding PIK is assigned. Table 1
presents the number of observations in each of the reference files.

A slightly confusing point: sometimes the Name and DOB reference files are described
as one and the same thing, and sometimes as separate.
I believe this is because **they differ only in how they are "cut" for the PVS process:**
the name reference file is cut by first and last initial,
while the DOB reference file is cut by month and day of birth.

This is described in Wagner and Layne, p.15:

> The [DOBSearch] module matches against a re-split
version of the Numident Name Reference file, splitting the data based on month and day
of birth.

Since we handle the logic of "cutting" in the linkage process itself, we generate
a single reference file here.

Note that unlike for addresses, and unlike for the pre-processing of *input* files
(as opposed to reference files), there is no explicit nickname processing/correction here.
I am fairly sure that is accurate to the PVS, which I believe assumes that nicknames
would not be present in SSA/tax records (or at least, that the real name would appear
at least once in these records).

(
    simulated_name_dob_numident_records,
    simulated_name_dob_numident_records_source_record_pairs,
) = df_ops.persist(
    merge_preserving_source_records(
        [simulated_alternate_name_numident, simulated_alternate_dob_numident],
        [
            simulated_alternate_name_numident_source_record_pairs,
            simulated_alternate_dob_numident_source_record_pairs,
        ],
        on=["ssn"],
        how="left",
        new_record_id_prefix="name_dob_numident_records",
    )
)
simulated_name_dob_numident_records

# Show alternate name/DOB for the same simulated SSN
simulated_name_dob_numident_records.merge(
    df_ops.groupby_agg_small_groups(
        simulated_name_dob_numident_records[["ssn"]],
        by="ssn",
        agg_func=lambda x: x.size(),
    )
    .pipe(lambda x: x[x > 1])
    .reset_index()[["ssn"]],
    on="ssn",
    how="inner",
).sort_values("ssn").tail(100)

#### Incorporating people with ITINs

Individual Taxpayer Identification Numbers (ITINs) can be issued to people who are required to file
federal taxes but are not eligible for a Social Security Number.
The most common reason for this is being an undocumented immigrant and therefore not being authorized
to work in the United States.

People without SSNs used to be impossible to assign PIKs to.
In 2011 the NORC report stated (p. 38, footnote 19):

> NORC understands that the Census Bureau has undertaken an effort to enhance the PVS reference files with IRS
files that include Individual Taxpayer Identification Numbers (ITIN). For those people who are required to file a tax
return but do not have, and may not want an SSN—such as a non-U.S. citizen—the IRS issues the taxpayer an ITIN.
This enhancement to the PVS reference file may help to match more non-U.S citizens.

By 2014 (Wagner and Layne, p. 5):

> One of the key enhancements [made in recent years] increased the coverage of the reference files by
including records for persons with Individual Taxpayer Identification Numbers assigned
by the Internal Revenue Service (ITINs) to [along with?] the SSN-based Numident data. 

I have not found a specific description of how ITIN records are constructed in any of the
publicly-available sources.
This may be because it is straightforward, or because the tax data schema is confidential.
I assume that only IRS data is used, since no other data source that I am aware of would
report ITIN.

It is stated that the ITIN records are created directly from tax filings and not
from ITIN applications (Brown et al. p. 29, footnote 16), which is convenient
because the tax filing data is what we can simulate with
pseudopeople:

> The NUMIDENT provides the PII on the SSN-holder from the issuing agency (SSA), and that PII is used in SSN
verification. **For ITINs, the Census Bureau does not have access to the ITIN applications** to the issuing agency (IRS),
so name and DOB verification of ITINs is less reliable.

"Less reliable" is a bit confusing here, because as stated above when generating
the simulated tax data, IRS data should not contain date of birth at all.
Here, we have stayed true to this by omitting it entirely.

We assume that only 1040 filings would be used for this purpose; we wouldn't expect ITINs to
show up on employer-filed W-2/1099 forms.

# Analogous to the process of getting alternate names and dates of birth
# from SSA, we retain all versions of the name from taxes.
simulated_name_for_itins, simulated_name_for_itins_source_record_pairs = df_ops.persist(
    dedupe_preserving_source_records(
        simulated_taxes_1040_with_itins.rename(
            columns={"middle_initial": "middle_name"}
        ),
        simulated_taxes_1040_with_itins_source_record_pairs,
        columns_to_dedupe=["ssn", "first_name", "middle_name", "last_name"],
        new_record_id_prefix="simulated_name_for_itins",
    )
)
simulated_name_for_itins

df_ops.compute(
    df_ops.groupby_agg_small_groups(
        simulated_name_for_itins, by="ssn", agg_func=lambda x: x.size()
    ).describe()
)

# How many records in the simulated name/DOB numident are in the ITIN range?
# With normal pseudopeople settings, this shouldn't happen, because SSN in SSA is noiseless.
# If this number is >0, we may be adding ITIN records as more alternates to existing records.
df_ops.compute(simulated_name_dob_numident_records.ssn.str.startswith("9").sum())

(
    simulated_name_dob_reference_file,
    simulated_name_dob_reference_file_source_record_pairs,
) = df_ops.persist(
    concat_preserving_source_records(
        [simulated_name_dob_numident_records, simulated_name_for_itins],
        [
            simulated_name_dob_numident_records_source_record_pairs,
            simulated_name_for_itins_source_record_pairs,
        ],
        new_record_id_prefix="simulated_name_dob_reference_file",
    )
)
simulated_name_dob_reference_file

### GeoBase Reference File

Wagner and Layne, p. 9:

> PVS creates three other sets of reference
files containing Numident data: the **GeoBase Reference File**, the Name Reference File,
and the DOB Reference file.
The GeoBase Reference File appends addresses from administrative records attached
to Numident data, including all possible combinations of alternate names and dates of
birth for SSN. Addresses from administrative records are edited and processed through
commercial software product to clean and standardize address data. ITIN data is also
incorporated into the Geobase.

Luque and Wagner, p. 5:

> Reference files contain data from the Numident file enhanced with address
data obtained from federal AR [administrative records] files.<sup>8</sup>
The reference files, thus, contain all variants of a person’s name, DOB,
and sex, as well as current and recent addresses. These reference files are
referred to as the (PVS) Geobase reference file since addresses (a geographic component)
are appended to each person record.<sup>9</sup> It is important to note that there are
multiple Geobase reference files that are created depending on the vintage of the
incoming file to be processed through PVS.

> <sup>8</sup> Namely, data from the IRS, Department of Housing and Urban Development,
several files from the Department of Health and Human Services, and Selective Service.

> <sup>9</sup> In particular, the address data is cleaned and standardized and used
to construct a variable called GEOKEY. The GEOKEY variable is constructed as a subset
of the full address, and then is appended to the Numident data to create the
PVS Geobase Reference file.

We only have IRS data to use for addresses, and specifically only W-2/1099 data,
which is a limitation of this case study.
I can't find a concrete definition of "recent" -- as noted above, we use 5 years
of IRS data.
This may be longer than the true window, but this may end up making up for
the lack of non-IRS data sources, and get us closer to a realistic number of
alternate addresses.

Also, our address data comes out of pseudopeople already parsed into address parts
like street name, etc.
For more realism, pseudopeople should output a single string that we have to (imperfectly) parse apart.

I haven't been able to find out more about what kind of "subset" the geokey is.
It is unclear to me why geokey is "interesting" since it is just derived from the
address parts.

address_cols = [c for c in simulated_taxes.columns if "mailing_address" in c]


def standardize_address_part(column):
    return (
        column
        # Remove leading or trailing whitespace
        .str.strip()
        # Turn any strings of consecutive whitespace into a single space
        .str.replace("\s+", " ", regex=True)
        # Normalize case
        .str.upper()
        # Normalize the word street as described in the example quoted above
        # In reality, there would be many rules like this
        .str.replace("\b(STREET|STR)\b", "ST", regex=True)
        # Make sure missingness is represented consistently
        .replace("", numpy.nan)
    )

simulated_tax_addresses = (
    simulated_taxes
    # Can only link these to the other files if they have an SSN
    .dropna(subset=["ssn"])[["record_id", "ssn"] + address_cols]
    .assign(
        **{c: lambda df, c=c: standardize_address_part(df[c]) for c in address_cols}
    )
    .pipe(df_ops.ensure_large_string_capacity)
)

simulated_addresses_by_ssn, simulated_addresses_by_ssn_source_record_pairs = (
    df_ops.persist(
        dedupe_preserving_source_records(
            simulated_tax_addresses,
            simulated_taxes_source_record_pairs,
            columns_to_dedupe=["ssn"] + address_cols,
            new_record_id_prefix="addresses_by_ssn",
        )
    )
)
simulated_addresses_by_ssn

num_addresses = (
    df_ops.groupby_agg_small_groups(
        simulated_addresses_by_ssn, by="ssn", agg_func=lambda x: x.size()
    )
    .rename("size")
    .reset_index()
    .sort_values("size")
)
num_addresses

# Show some SSNs with a lot of address variation
df_ops.compute(
    pd.merge(
        num_addresses.tail(10), simulated_addresses_by_ssn, on="ssn", how="inner"
    ).sort_values("ssn")
)

# Rough estimate of how many rows we should have in our reference file, once we do this Cartesian product
df_ops.compute(
    len(simulated_name_dob_reference_file)
    * df_ops.groupby_agg_small_groups(
        simulated_addresses_by_ssn, by="ssn", agg_func=lambda x: x.size()
    ).mean()
)

(
    simulated_geobase_reference_file,
    simulated_geobase_reference_file_source_record_pairs,
) = df_ops.persist(
    merge_preserving_source_records(
        [simulated_name_dob_reference_file, simulated_addresses_by_ssn],
        [
            simulated_name_dob_reference_file_source_record_pairs,
            simulated_addresses_by_ssn_source_record_pairs,
        ],
        on=["ssn"],
        how="left",
        new_record_id_prefix="simulated_geobase_reference_file",
    )
)
simulated_geobase_reference_file

# Actual number of rows
len(simulated_geobase_reference_file)

## Track ground truth for reference files

def get_simulants_of_source_records(source_record_pairs, filter_record_ids=None):
    if filter_record_ids is not None:
        source_record_pairs = source_record_pairs.pipe(filter_record_ids)

    result = (
        df_ops.drop_duplicates(source_record_pairs)
        .merge(source_record_ground_truth, on="source_record_id")[
            ["record_id", "simulant_id"]
        ]
        .pipe(df_ops.drop_duplicates)
    )
    result = result.merge(
        df_ops.groupby_agg_small_groups(
            result,
            by="record_id",
            agg_func=lambda x: x.simulant_id.nunique().rename("n_unique_simulants"),
        ).reset_index(),
        on="record_id",
        how="left",
    )
    return result


def get_ground_truth_for_records(source_record_pairs):
    result = df_ops.persist(get_simulants_of_source_records(source_record_pairs))

    result_nunique_describe = df_ops.compute(result["n_unique_simulants"].describe())

    if numpy.isclose(result_nunique_describe.loc["max"], 1):
        print("No collisions")
        return result

    print("Collisions:")
    display(result_nunique_describe)

    print("Simulated tax records with the most collisions:")
    most_collisions_record_id = (
        result.sort_values("n_unique_simulants", ascending=False)
        .head(1)
        .record_id.iloc[0]
    )
    most_collisions_source_record_ids = df_ops.compute(
        df_ops.drop_duplicates(
            source_record_pairs[
                source_record_pairs.record_id == most_collisions_record_id
            ][["source_record_id"]]
        )
    )
    tax_record_ids = most_collisions_source_record_ids.merge(
        df_ops.compute(
            df_ops.drop_duplicates(
                simulated_taxes_source_record_pairs[
                    simulated_taxes_source_record_pairs.source_record_id.isin(
                        most_collisions_source_record_ids.source_record_id
                    )
                ][["record_id", "source_record_id"]]
            )
        ),
        on="source_record_id",
        how="inner",
    )
    most_collisions_ground_truth = df_ops.compute(
        source_record_ground_truth[
            source_record_ground_truth.source_record_id.isin(
                most_collisions_source_record_ids.source_record_id
            )
        ]
    )

    most_collisions_tax_filings = tax_record_ids.merge(
        df_ops.compute(
            simulated_taxes[simulated_taxes.record_id.isin(tax_record_ids.record_id)],
        ),
        on="record_id",
        how="left",
    ).merge(
        most_collisions_ground_truth,
        on="source_record_id",
        how="left",
    )
    display(most_collisions_tax_filings)

    print("Simulated SSA records with the most collisions:")
    display(
        df_ops.compute(
            most_collisions_source_record_ids.merge(
                df_ops.compute(
                    simulated_ssa_numident[
                        simulated_ssa_numident.record_id.isin(
                            most_collisions_source_record_ids.source_record_id
                        )
                    ].rename(columns={"record_id": "source_record_id"})
                ),
                on="source_record_id",
                how="inner",
            ).merge(most_collisions_ground_truth, on="source_record_id", how="left")
        )
    )

    return result

### Simulated Census Numident

len(simulated_census_numident_source_record_pairs)

simulated_census_numident_ground_truth = get_ground_truth_for_records(
    simulated_census_numident_source_record_pairs
)

simulated_census_numident_ground_truth

### Simulated Alternate Name Numident

len(simulated_alternate_name_numident_source_record_pairs)

simulated_alternate_name_numident_ground_truth = get_ground_truth_for_records(
    simulated_alternate_name_numident_source_record_pairs
)

simulated_alternate_name_numident_ground_truth

### Alternate DOB Numident

len(simulated_alternate_dob_numident_source_record_pairs)

simulated_alternate_dob_numident_ground_truth = get_ground_truth_for_records(
    simulated_alternate_dob_numident_source_record_pairs
)

simulated_alternate_dob_numident_ground_truth

### Simulated Name/DOB Reference File

len(simulated_name_dob_reference_file_source_record_pairs)

simulated_name_dob_reference_file_ground_truth = get_ground_truth_for_records(
    simulated_name_dob_reference_file_source_record_pairs
)

simulated_name_dob_reference_file_ground_truth

### Simulated GeoBase Reference File

len(simulated_geobase_reference_file_source_record_pairs)

simulated_geobase_reference_file_ground_truth = get_ground_truth_for_records(
    simulated_geobase_reference_file_source_record_pairs
)

simulated_geobase_reference_file_ground_truth

%xdel simulated_taxes
%xdel simulated_taxes_source_record_pairs

### Get ground truth by SSN

simulated_ssn_simulant_pairs = df_ops.persist(
    df_ops.concat(
        [
            simulated_census_numident[["record_id", "ssn"]].merge(
                simulated_census_numident_ground_truth, on="record_id"
            ),
            simulated_alternate_name_numident[["record_id", "ssn"]].merge(
                simulated_alternate_name_numident_ground_truth, on="record_id"
            ),
            simulated_alternate_dob_numident[["record_id", "ssn"]].merge(
                simulated_alternate_dob_numident_ground_truth, on="record_id"
            ),
            simulated_name_dob_reference_file[["record_id", "ssn"]].merge(
                simulated_name_dob_reference_file_ground_truth, on="record_id"
            ),
            simulated_geobase_reference_file[["record_id", "ssn"]].merge(
                simulated_geobase_reference_file_ground_truth, on="record_id"
            ),
        ]
    )[["ssn", "simulant_id"]].pipe(df_ops.drop_duplicates)
)
simulated_ssn_simulant_pairs

# How many simulated SSNs have different simulant IDs
# contributing to them?
df_ops.compute(
    (
        df_ops.groupby_agg_small_groups(
            simulated_ssn_simulant_pairs,
            by="ssn",
            agg_func=lambda x: x.simulant_id.nunique(),
        )
        > 1
    ).sum()
)

## Save results

files = {
    "simulated_census_numident": (
        simulated_census_numident,
        simulated_census_numident_ground_truth,
    ),
    "simulated_alternate_name_numident": (
        simulated_alternate_name_numident,
        simulated_alternate_name_numident_ground_truth,
    ),
    "simulated_alternate_dob_numident": (
        simulated_alternate_dob_numident,
        simulated_alternate_dob_numident_ground_truth,
    ),
    "simulated_geobase_reference_file": (
        simulated_geobase_reference_file,
        simulated_geobase_reference_file_ground_truth,
    ),
    "simulated_name_dob_reference_file": (
        simulated_name_dob_reference_file,
        simulated_name_dob_reference_file_ground_truth,
    ),
}

simulated_reference_files = [
    simulated_census_numident,
    simulated_alternate_name_numident,
    simulated_alternate_dob_numident,
    simulated_geobase_reference_file,
    simulated_name_dob_reference_file,
]
# TODO: Rename the ssn column to explicitly include itins, since this is confusing
simulated_all_ssns_itins_in_reference_files = df_ops.concat(
    [df[["ssn"]] for df in simulated_reference_files], ignore_index=True
)
simulated_ssn_to_pik = simulated_all_ssns_itins_in_reference_files.pipe(
    df_ops.drop_duplicates
).pipe(add_unique_id_col, col_name="pik")[["ssn", "pik"]]
simulated_ssn_to_pik

simulated_pik_simulant_pairs = simulated_ssn_simulant_pairs.merge(
    simulated_ssn_to_pik, on="ssn", how="inner"
)[["pik", "simulant_id"]]
simulated_pik_simulant_pairs

%xdel simulated_ssn_simulant_pairs

for file_name, (file, ground_truth) in files.items():
    file = df_ops.persist(file.merge(simulated_ssn_to_pik, on="ssn", how="left"))
    assert df_ops.compute(file.pik.notnull().all())

    save_file_with_ground_truth(file_name, file, ground_truth)

simulated_pik_simulant_pairs_path = (
    f"{output_dir}/{data_to_use}/simulated_pik_simulant_pairs.parquet"
)
utils.remove_path(simulated_pik_simulant_pairs_path)
simulated_pik_simulant_pairs.to_parquet(simulated_pik_simulant_pairs_path)

print(datetime.datetime.now())
